Move log levels to Parity tab
