+++ js/src/views/Settings/index.js
@@ -16,7 +16,6 @@
 
 import settingsReducer from './reducers';
 import { toggleView, updateBackground } from './actions';
-import SettingsAdvanced from './Advanced';
 import SettingsBackground from './Background';
 import SettingsParity from './Parity';
 import SettingsProxy from './Proxy';
@@ -25,7 +24,6 @@ import SettingsViews, { defaultViews } from './Views';
 export default from './settings';
 
 export {
-  SettingsAdvanced,
   SettingsBackground,
   SettingsParity,
   SettingsProxy,
